services/core/requirements.txt
Django>=6.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-filter
python-json-logger
sentry-sdk
celery
redis
python-decouple
Pillow
cryptography
stripe

# Testing
pytest
pytest-django
factory-boy
freezegun
websockets
httpx
pytest-asyncio
